Add WordPressReader target
