+++ Modules/Package.swift
@@ -14,6 +14,7 @@ let package = Package(
         .library(name: "WordPressFlux", targets: ["WordPressFlux"]),
         .library(name: "WordPressShared", targets: ["WordPressShared"]),
         .library(name: "WordPressUI", targets: ["WordPressUI"]),
+        .library(name: "WordPressReader", targets: ["WordPressReader"]),
     ],
     dependencies: [
         .package(url: "https://github.com/airbnb/lottie-ios", from: "4.4.0"),
@@ -69,12 +70,13 @@ let package = Package(
         .target(name: "WordPressTesting", resources: [.process("Resources")]),
         .target(
             name: "WordPressUI",
-            dependencies: [
-                "AsyncImageKit",
-                .target(name: "WordPressShared")
-            ],
+            dependencies: ["AsyncImageKit", "WordPressShared"],
             resources: [.process("Resources")],
             swiftSettings: [.swiftLanguageMode(.v5)]
+        ),
+        .target(
+            name: "WordPressReader",
+            dependencies: ["AsyncImageKit", "WordPressUI", "WordPressShared"]
         ),
         .testTarget(name: "JetpackStatsWidgetsCoreTests", dependencies: [.target(name: "JetpackStatsWidgetsCore")], swiftSettings: [.swiftLanguageMode(.v5)]),
         .testTarget(name: "DesignSystemTests", dependencies: [.target(name: "DesignSystem")], swiftSettings: [.swiftLanguageMode(.v5)]),
@@ -156,6 +158,7 @@ enum XcodeSupport {
                 "JetpackStatsWidgetsCore",
                 "WordPressFlux",
                 "WordPressShared",
+                "WordPressReader",
                 "AsyncImageKit",
                 "WordPressUI",
                 "WordPressCore",
