+++ WordPress/Classes/ViewRelated/Reader/Cards/Tags View/ReaderTopicCollectionViewCoordinator.swift
@@ -1,4 +1,5 @@
 import UIKit
+import WordPressReader
 
 enum ReaderTopicCollectionViewState {
     case collapsed
