+++ WordPress/Classes/ViewRelated/Reader/Detail/ReaderDetailCoordinator.swift
@@ -1,5 +1,6 @@
 import Foundation
 import WordPressShared
+import WordPressReader
 import AsyncImageKit
 import Combine
