+++ WordPress/Classes/ViewRelated/Reader/Detail/Views/ReaderDetailHeaderView.swift
@@ -1,6 +1,7 @@
 import SwiftUI
 import AsyncImageKit
 import WordPressUI
+import WordPressReader
 
 protocol ReaderDetailHeaderViewDelegate: AnyObject {
     func didTapBlogName()
