+++ WordPress/Classes/ViewRelated/Reader/ReaderFollowButton.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import WordPressReader
 
 struct ReaderFollowButton: View {
