+++ WordPress/Classes/ViewRelated/Reader/Theme/ReaderDisplaySettingViewController.swift
@@ -1,5 +1,6 @@
 import SwiftUI
 import DesignSystem
+import WordPressReader
 
 /// The tracking source values for the customization sheet.
 /// The values are kept in sync with Android.
